# Fix delay for negative steps and optimize perfomance